# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Bootcamp Management System APIs",
    "version": "1.0",
    "description": "Bootcamp Management System APIs",
    "contact": {
      "name": "VMware "
    }
  },
  "servers": [
    {
      "url": "https://localhost:4000"
    }
  ],
  "paths": {
...
    "/api/assignments": {
      "get": {
        "summary": "Get all assignments",
        "tags": [
          "AssignmentAPIs"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "message": "Get request to all Team assignment ",
                      "assignmentList": [
                        {
                          "_id": "62dd4695845e256444ee1ac4",
                          "assignmentName": "Db assignment",
                          "assignmentDescriptions": "database assignment",
                          "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                          "postingDate": "2022-07-24T13:18:13.530Z",
                          "deadlineDate": "2022-08-09T18:30:00.000Z",
                          "isProject": false,
                          "isTeamAssignment": true,
                          "maximumMarks": 100,
                          "grading": "Absolute",
                          "__v": 0
                        }
                      ]
                    }
                  },
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "assignmentList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "assignmentName": {
                            "type": "string"
                          },
                          "assignmentDescriptions": {
                            "type": "string"
                          },
                          "assignmentRepositoryUrl": {
                            "type": "string"
                          },
                          "postingDate": {
                            "type": "string"
                          },
                          "deadlineDate": {
                            "type": "string"
                          },
                          "isProject": {
                            "type": "boolean"
                          },
                          "isTeamAssignment": {
                            "type": "boolean"
                          },
                          "maximumMarks": {
                            "type": "integer"
                          },
                          "grading": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "Example (based on query parameter)": {
                    "value": {
                      "message": "Get request to all Team assignment ",
                      "assignmentList": [
                        {
                          "_id": "62dd4695845e256444ee1ac4",
                          "assignmentName": "Db assignment",
                          "assignmentDescriptions": "database assignment",
                          "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                          "postingDate": "2022-07-24T13:18:13.530Z",
                          "deadlineDate": "2022-08-09T18:30:00.000Z",
                          "isProject": false,
                          "isTeamAssignment": true,
                          "maximumMarks": 100,
                          "grading": "Absolute",
                          "__v": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "operationId": "get-assignments",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "all",
            "description": "retrieves all assignments"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "project",
            "description": "retrieves assignments that are projects"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "individual",
            "description": "retrieves individual assignments"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "team",
            "description": "retrieves team assignments"
          }
        ],
        "description": "Get all assignments"
      },
      "post": {
        "summary": "Post Assignment",
        "operationId": "post-assignment",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "assignmentName": {
                      "type": "string"
                    },
                    "assignmentDescriptions": {
                      "type": "string"
                    },
                    "assignmentRepositoryUrl": {
                      "type": "string"
                    },
                    "postingDate": {
                      "type": "string"
                    },
                    "deadlineDate": {
                      "type": "string"
                    },
                    "isProject": {
                      "type": "boolean"
                    },
                    "isTeamAssignment": {
                      "type": "boolean"
                    },
                    "maximumMarks": {
                      "type": "integer"
                    },
                    "grading": {
                      "type": "string"
                    },
                    "_id": {
                      "type": "string"
                    },
                    "__v": {
                      "type": "integer"
                    }
                  },
                  "x-examples": {
                    "example-1": {
                      "assignmentName": "Db assignment",
                      "assignmentDescriptions": "database assignment",
                      "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                      "postingDate": "2022-07-24T13:27:35.604Z",
                      "deadlineDate": "2022-08-09T18:30:00.000Z",
                      "isProject": false,
                      "isTeamAssignment": true,
                      "maximumMarks": 100,
                      "grading": "Absolute",
                      "_id": "62dd48c7845e256444ee1acb",
                      "__v": 0
                    }
                  }
                },
                "examples": {
                  "Response body example": {
                    "value": {
                      "assignmentName": "Db assignment",
                      "assignmentDescriptions": "database assignment",
                      "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                      "postingDate": "2022-07-24T13:27:35.604Z",
                      "deadlineDate": "2022-08-09T18:30:00.000Z",
                      "isProject": false,
                      "isTeamAssignment": true,
                      "maximumMarks": 100,
                      "grading": "Absolute",
                      "_id": "62dd48c7845e256444ee1acb",
                      "__v": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "tags": [
          "AssignmentAPIs"
        ],
        "description": "Create a new assignment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isProject": {
                    "type": "boolean"
                  },
                  "assignmentName": {
                    "type": "string"
                  },
                  "assignmentDescriptions": {
                    "type": "string"
                  },
                  "assignmentRepositoryUrl": {
                    "type": "string"
                  },
                  "deadlineDate": {
                    "type": "string"
                  },
                  "isTeamAssignment": {
                    "type": "boolean"
                  },
                  "grading": {
                    "type": "string"
                  },
                  "maximumMarks": {
                    "type": "integer"
                  }
                },
                "x-examples": {
                  "example-1": {
                    "isProject": false,
                    "assignmentName": "Db assignment",
                    "assignmentDescriptions": "database assignment",
                    "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                    "deadlineDate": "2022/08/10",
                    "isTeamAssignment": true,
                    "grading": "Absolute",
                    "maximumMarks": 100
                  }
                }
              },
              "examples": {
                "Request body Example": {
                  "value": {
                    "isProject": false,
                    "assignmentName": "Db assignment",
                    "assignmentDescriptions": "database assignment",
                    "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                    "deadlineDate": "2022/08/10",
                    "isTeamAssignment": true,
                    "grading": "Absolute",
                    "maximumMarks": 100
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/assignment/{assignmentId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "assignmentId",
          "in": "path",
          "required": true,
          "description": "Assignment Id"
        }
      ],
      "patch": {
        "summary": "Patch Assignment by Id",
        "tags": [
          "AssignmentAPIs"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "updatedAssignmentObject": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "assignmentName": {
                          "type": "string"
                        },
                        "assignmentDescriptions": {
                          "type": "string"
                        },
                        "assignmentRepositoryUrl": {
                          "type": "string"
                        },
                        "postingDate": {
                          "type": "string"
                        },
                        "deadlineDate": {
                          "type": "string"
                        },
                        "isProject": {
                          "type": "boolean"
                        },
                        "isTeamAssignment": {
                          "type": "boolean"
                        },
                        "maximumMarks": {
                          "type": "integer"
                        },
                        "grading": {
                          "type": "string"
                        },
                        "__v": {
                          "type": "integer"
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "example-1": {
                      "message": "Assignment updated",
                      "updatedAssignmentObject": {
                        "_id": "62dd4695845e256444ee1ac4",
                        "assignmentName": "Db assignment",
                        "assignmentDescriptions": "database assignment",
                        "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                        "postingDate": "2022-07-24T13:18:13.530Z",
                        "deadlineDate": "2022-08-09T18:30:00.000Z",
                        "isProject": false,
                        "isTeamAssignment": true,
                        "maximumMarks": 100,
                        "grading": "Absolute",
                        "__v": 0
                      }
                    }
                  }
                },
                "examples": {
                  "Response body example": {
                    "value": {
                      "message": "Assignment updated",
                      "updatedAssignmentObject": {
                        "_id": "62dd4695845e256444ee1ac4",
                        "assignmentName": "Db assignment",
                        "assignmentDescriptions": "database assignment",
                        "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                        "postingDate": "2022-07-24T13:18:13.530Z",
                        "deadlineDate": "2022-08-09T18:30:00.000Z",
                        "isProject": false,
                        "isTeamAssignment": true,
                        "maximumMarks": 100,
                        "grading": "Absolute",
                        "__v": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "x-examples": {
                    "example-1": {
                      "message": "There is no such Assignment"
                    }
                  }
                },
                "examples": {
                  "Error message ": {
                    "value": {
                      "message": "There is no such Assignment"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "update-assignment",
        "description": "Update an assignment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "x-examples": {
                  "example-1": {
                    "isProject": true,
                    "isTeamAssignment": false
                  }
                }
              },
              "examples": {
                "Request body example": {
                  "value": {
                    "isProject": true,
                    "isTeamAssignment": true
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Assignment by Id",
        "tags": [
          "AssignmentAPIs"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "deletedAssignmentObject": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "assignmentName": {
                          "type": "string"
                        },
                        "assignmentDescriptions": {
                          "type": "string"
                        },
                        "assignmentRepositoryUrl": {
                          "type": "string"
                        },
                        "postingDate": {
                          "type": "string"
                        },
                        "deadlineDate": {
                          "type": "string"
                        },
                        "isProject": {
                          "type": "boolean"
                        },
                        "isTeamAssignment": {
                          "type": "boolean"
                        },
                        "maximumMarks": {
                          "type": "integer"
                        },
                        "grading": {
                          "type": "string"
                        },
                        "__v": {
                          "type": "integer"
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "example-1": {
                      "message": "Successfully Deleted Assignment",
                      "deletedAssignmentObject": {
                        "_id": "62dd48c7845e256444ee1acb",
                        "assignmentName": "Db assignment",
                        "assignmentDescriptions": "database assignment",
                        "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                        "postingDate": "2022-07-24T13:27:35.604Z",
                        "deadlineDate": "2022-08-09T18:30:00.000Z",
                        "isProject": false,
                        "isTeamAssignment": true,
                        "maximumMarks": 100,
                        "grading": "Absolute",
                        "__v": 0
                      }
                    }
                  }
                },
                "examples": {
                  "Example": {
                    "value": {
                      "message": "Successfully Deleted Assignment",
                      "deletedAssignmentObject": {
                        "_id": "62dd48c7845e256444ee1acb",
                        "assignmentName": "Db assignment",
                        "assignmentDescriptions": "database assignment",
                        "assignmentRepositoryUrl": "https://gitlab.eng.vmware.com/user/test",
                        "postingDate": "2022-07-24T13:27:35.604Z",
                        "deadlineDate": "2022-08-09T18:30:00.000Z",
                        "isProject": false,
                        "isTeamAssignment": true,
                        "maximumMarks": 100,
                        "grading": "Absolute",
                        "__v": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "x-examples": {
                    "example-1": {
                      "message": "There is no such Assignment"
                    }
                  }
                },
                "examples": {
                  "Error message": {
                    "value": {
                      "message": "There is no such Assignment"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "delete-assignment",
        "description": "Delete an assignment"
      }
    },
...
  }
}